More impl refactoring
