+++ backend/src/nodes/impl/fill_alpha.py
@@ -4,8 +4,7 @@ import math
 import cv2
 import numpy as np
 
-from .blend_modes import BlendModes
-from .image_utils import blend_images
+from .blend import BlendModes, blend_images
 from ..utils.utils import get_h_w_c
