+++ backend/src/nodes/impl/normals/util.py
@@ -1,0 +1,20 @@
+from typing import Tuple
+import numpy as np
+
+
+def normalize_normals(
+    x: np.ndarray, y: np.ndarray
+) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
+    # The square of the length of X and Y
+    l_sq = np.square(x) + np.square(y)
+
+    # If the length of X and Y is >1, then make it 1
+    l = np.sqrt(np.maximum(l_sq, 1))
+    x /= l
+    y /= l
+    l_sq = np.minimum(l_sq, 1, out=l_sq)
+
+    # Compute Z
+    z = np.sqrt(1 - l_sq)
+
+    return x, y, z
