+++ backend/src/nodes/nodes/image_filter/normal_addition.py
@@ -9,7 +9,7 @@ from ...node_factory import NodeFactory
 from ...properties.inputs import ImageInput, SliderInput
 from ...properties.outputs import ImageOutput
 from ...properties import expression
-from ...impl.image_utils import normalize_normals
+from ...impl.normals.util import normalize_normals
 
 
 @NodeFactory.register("chainner:image:add_normals")
