+++ backend/src/nodes/nodes/image_utility/blend.py
@@ -9,10 +9,8 @@ from ...node_factory import NodeFactory
 from ...properties.inputs import ImageInput, BlendModeDropdown
 from ...properties.outputs import ImageOutput
 from ...properties import expression
-from ...impl.image_utils import (
-    as_2d_grayscale,
-    blend_images,
-)
+from ...impl.blend import blend_images
+from ...impl.image_utils import as_2d_grayscale
 from ...impl.pil_utils import convert_to_BGRA
 from ...utils.utils import get_h_w_c
