+++ backend/src/nodes/nodes/image_utility/caption.py
@@ -12,7 +12,7 @@ from ...properties.inputs import (
     CaptionPositionInput,
 )
 from ...properties.outputs import ImageOutput
-from ...impl.pil_utils import add_caption
+from ...impl.caption import add_caption
 
 
 @NodeFactory.register("chainner:image:caption")
