+++ backend/src/nodes/properties/inputs/file_inputs.py
@@ -4,7 +4,7 @@ from typing import Literal, Union
 import os
 
 # pylint: disable=relative-beyond-top-level
-from ...impl.image_utils import get_available_image_formats
+from ...impl.image_formats import get_available_image_formats
 from .. import expression
 from .base_input import BaseInput
 from .generic_inputs import DropDownInput
